Move buildElasticAggregationsBody
