+++ lib/elasticsearch/elastic-body-builder.js
@@ -1,4 +1,4 @@
-const { EXCLUDE_FIELDS, ITEM_FILTER_AGGREGATIONS, SORT_FIELDS } = require('./config')
+const { EXCLUDE_FIELDS, ITEM_FILTER_AGGREGATIONS, SORT_FIELDS, AGGREGATIONS_SPEC } = require('./config')
 const { deepValue } = require('../util')
 const { esRangeValue } = require('../utils/resource-helpers')
 const ApiRequest = require('../api-request')
@@ -380,6 +380,26 @@ const bodyForSearch = function (params) {
   return body
 }
 
+const buildElasticAggregationsBody = (params, aggregateProps) => {
+  // Add an `aggregations` entry to the ES body describing the aggretations
+  // we want. Set the `size` property to per_page (default 50) for each.
+  // https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-terms-aggregation.html#search-aggregations-bucket-terms-aggregation-size
+  const aggregations = aggregateProps.reduce((aggs, prop) => {
+    aggs[prop] = AGGREGATIONS_SPEC[prop]
+    // Only set size for terms aggs for now:
+    if (aggs[prop].terms) {
+      aggs[prop].terms.size = params.per_page
+    }
+    return aggs
+  }, {})
+
+  const body = buildElasticBody(params)
+  body.size = 0
+  body.aggregations = aggregations
+
+  return body
+}
+
 module.exports = {
   bodyForFindByUri,
   addInnerHits,
@@ -389,5 +409,6 @@ module.exports = {
   itemsQueryContext,
   buildElasticQuery,
   buildElasticBody,
-  bodyForSearch
+  bodyForSearch,
+  buildElasticAggregationsBody
 }
