+++ lib/resources.js
@@ -26,7 +26,8 @@ const {
   itemsQueryContext,
   buildElasticQuery,
   buildElasticBody,
-  bodyForSearch
+  bodyForSearch,
+  buildElasticAggregationsBody
 } = require('./elasticsearch/elastic-body-builder')
 
 const RESOURCES_INDEX = process.env.RESOURCES_INDEX
@@ -205,26 +206,6 @@ module.exports = function (app, _private = null) {
         }
         throw new IndexSearchError(`Error processing search: ${e.message || e}`)
       })
-  }
-
-  const buildElasticAggregationsBody = (params, aggregateProps) => {
-    // Add an `aggregations` entry to the ES body describing the aggretations
-    // we want. Set the `size` property to per_page (default 50) for each.
-    // https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-terms-aggregation.html#search-aggregations-bucket-terms-aggregation-size
-    const aggregations = aggregateProps.reduce((aggs, prop) => {
-      aggs[prop] = AGGREGATIONS_SPEC[prop]
-      // Only set size for terms aggs for now:
-      if (aggs[prop].terms) {
-        aggs[prop].terms.size = params.per_page
-      }
-      return aggs
-    }, {})
-
-    const body = buildElasticBody(params)
-    body.size = 0
-    body.aggregations = aggregations
-
-    return body
   }
 
   /**
